feature/ON4T-173_TaskCollection_Backend Fixed Unit Tests and added new Unit Tests

+++ src/main/java/com/gocaspi/taskfly/taskcollection/TaskCollectionService.java
@@ -52,6 +52,9 @@ public class TaskCollectionService {
 
 
     public void deleteTaskCollectionByID(String id){
+        if (!repo.existsById(id)){
+            throw httpNotFoundError;
+        }
         repo.deleteById(id);
     }
